removing bot reaction

+++ src/bot/eventListeners/messageReactions.ts
@@ -6,6 +6,7 @@ import {
 } from '../../utils/sparkleUtils.js'
 import { loadConfigData } from '../services/configService.js'
 import { ConfigData } from '../../types.js'
+import discordClient from '../../config/discordConfig.js'
 
 // TODOLATER: check for user permissions before proceeding
 // TODOLATER: decide if we want to delete the message or just remove the reaction
@@ -134,7 +135,33 @@ export default (client: Client) => {
     try {
       if (reaction.emoji.name !== '✨') return
 
+      const authorized = authorizeUserSparkle(reaction)
+
+      if (!authorized) {
+        console.log(`User is not authorized to manage messages`)
+        return
+      }
+
       await deleteSparkleMessage(reaction, user)
+      const replyToRemoveSparkle = async (reaction) => {
+        try {
+          // Iterate over all reactions on the message
+          const botReactions = reaction.message.reactions.cache.filter((r) =>
+            r.users.cache.has(discordClient.user.id)
+          )
+
+          for (const botReaction of botReactions.values()) {
+            // Remove the bot's reaction
+            await botReaction.users.remove(discordClient.user.id)
+            console.log(`Removed bot's ${botReaction.emoji.name} reaction`)
+          }
+        } catch (error) {
+          console.error('Error removing bot reactions:', error)
+          return { error: 'Failed to remove bot reactions.' }
+        }
+      }
+
+      await replyToRemoveSparkle(reaction)
     } catch (error) {
       console.error('Error handling messageReactionRemove event:', error)
     }
